Refactored RoutesWrapper router component
