+++ src/router/components/RoutesWrapper/index.ts
@@ -1,0 +1,2 @@
+import RoutesWrapper from './RoutesWrapper'
+export default RoutesWrapper
